// src/api/models/dto/UserResponse.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type DtoUserResponse = {
    /**
     * @type string | undefined
    */
    avatar_url?: string;
    /**
     * @type string | undefined
    */
    bio?: string;
    /**
     * @type string | undefined
    */
    cover_url?: string;
    /**
     * @type string | undefined
    */
    created_at?: string;
    /**
     * @type string | undefined
    */
    display_name?: string;
    /**
     * @type string | undefined
    */
    email?: string;
    /**
     * @type integer | undefined
    */
    follower_count?: number;
    /**
     * @type integer | undefined
    */
    following_count?: number;
    /**
     * @type string | undefined
    */
    id?: string;
    /**
     * @type boolean | undefined
    */
    is_active?: boolean;
    /**
     * @type boolean | undefined
    */
    is_following?: boolean;
    /**
     * @type string | undefined
    */
    location?: string;
    /**
     * @type string | undefined
    */
    updated_at?: string;
    /**
     * @type string | undefined
    */
    username?: string;
    /**
     * @type string | undefined
    */
    website?: string;
};